delete old code for Cartitem system

+++ product/models.py
@@ -29,7 +29,7 @@ class City(models.Model):
 class Product(models.Model):
     name = models.CharField(max_length=255)
     description = models.TextField()
-    price = models.FloatField()
+    price = models.DecimalField(decimal_places=2, max_digits=7)
     image = models.ImageField(upload_to="products/")
     category = models.ForeignKey(
         Category, on_delete=models.CASCADE, null=True, blank=True
